requirements.txt
azure-core==1.2.1
azure-storage-blob==12.1.0
azure-storage-file-datalake==12.0.0b6
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
idna==2.8
isodate==0.6.0
msrest==0.6.10
oauthlib==3.1.0
pycparser==2.19
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
urllib3==1.25.7
